Fixing 1 in 256 edge-case that can lead to byte shift followed by checksum mismatch and therefore no output.
If the last byte of the checksum is 0x42 it can be mistaken for the first start-byte, therefore checking the secondy start-byte is neccessary.
The probability of this happening is 1 in 256 and in addition it has 100% probability if testing hepa-filtered air with 0 particels.
Reason: All non-zero bytes remaining are: 
Decimal: 66 + 77 + 28 + 151 checksum: 256 + 66 
Hex: 0x42 0x4d 0x00 0x1c ... 0x97 checksum: 0x01 0x42

It is not possible for byte 31 (second checksum-byte) to be 0x42 too, as the sum of the first 30 bytes is at most 30*255 which is less than 66 * 256 which is the smallest checksum that would cause byte 31 to be 0x42 (=66). The sequence 0x42 0x42 0x42 0x4d with the last two being the start bytes is therefore impossible.

Though it is not impossible for the sequence 0x42 0x42 0x4d or 0x42 0x4d to appear at another position in the sequence it can only happen with particle count above 16896 = 66*256 and then this event is extremely unlikely itself and will extremely likely lead to a checksum mismatch therefore be discarded.